Support decimal manga chapters and harden Mongo/local backup
Manga sites commonly use fractional chapter numbers (2.1, 12.5). Progress
was parseInt'd everywhere, so those chapters collapsed to the integer and
could not be stored, listed, or used as "next chapter".

Progress is now a float through the sync page, episode list, number input,
and storage providers. Integer-only trackers (AniList, MAL, Kitsu,
Shikimori) floor at the API boundary, and once they hold N we do not
re-prompt for N.x. SpaceTimeDB stores progress/volumeProgress as f64.

Site parsers (chibi + classic pages) capture optional decimals instead of
stripping them with /\D+/ or parseInt. Next-chapter lookup walks the
episode map for the smallest key above the current number, not always +1.

Also makes MongoDB and local lists usable when the backend is a LAN or
localhost server, or briefly offline:

- Route MongoDB requests through the background xhr helper so content
  scripts are not blocked by mixed-content / private-network access.
- Retry GET failures and surface ServerOfflineError instead of a generic
  fetch crash.
- Request optional host permission when applying a custom MongoDB URL
  (new settings control). Allow localhost/127.0.0.1 in the manifest.
- Write-through local:// backup on MongoDB/SpaceTimeDB sync so progress
  is kept on-device even if the DB write fails. Backup is write-only and
  never read back into the active provider automatically.
- Bookmarks list falls back to Local when the DB is unreachable, and the
  Local <-> DB buttons work for both MongoDB and SpaceTimeDB (disabled
  while offline).
- Sync page errors are caught and shown instead of becoming uncaught
  rejections (including Mangadex lifecycle).

Misc:
- Deduplicate custom-domain content-script registrations.
- Drop Firefox-only manifest keys from Chrome/unpacked builds.
- Remove duplicate AsuraScans match patterns.
- Log "Background list Sync not allowed" as info, not error.

NOTE: SpaceTimeDB schema changed (progress/volumeProgress u32 -> f64).
Restart/republish the module before writing new entries.

+++ spacetimedb/src/schema.ts
@@ -36,8 +36,9 @@ export const SyncEntry = table(
 		image: t.string().optional(),
 		tags: t.string(),
 		streamingUrl: t.string().optional(),
-		progress: t.u32(),
-		volumeProgress: t.u32(),
+		// f64 so manga chapter progress can be fractional (2.1, 2.2, …).
+		progress: t.f64(),
+		volumeProgress: t.f64(),
 		score: t.u8(),
 		status: t.u8(),
 		updatedAt: t.timestamp(),
